Fixed type exports
