+++ src/index.tsx
@@ -1,4 +1,10 @@
 import ReactNativeZoomableView from './ReactNativeZoomableView';
 import ReactNativeZoomableViewWithGestures from './ReactNativeZoomableViewWithGestures';
+import { ReactNativeZoomableViewProps, ZoomableViewEvent } from './typings';
 
-export { ReactNativeZoomableView, ReactNativeZoomableViewWithGestures };
+export {
+  ReactNativeZoomableView,
+  ReactNativeZoomableViewWithGestures,
+  ReactNativeZoomableViewProps,
+  ZoomableViewEvent,
+};
